feat: Consumir API de ml-ia completo RIA-01

+++ backend/src/main/java/com/pipre/backend/domain/Prediction.java
@@ -1,11 +1,12 @@
 package com.pipre.backend.domain;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public record Prediction(
         String result,
-        Float accuracy,
-        Float precision,
+        BigDecimal accuracy,
+        BigDecimal precision,
         Map<String, Object> details
 ) {
 }
